split TodoForm class component into separate file, import to App
